+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 :warning: This server code is versioned separately to the download of the [Hiro game framework](https://heroiclabs.com/hiro/). :warning:
 
 ## [Unreleased]
+### Added
+- Add switches for core and authenticate events to be sent by the SatoriPersonalizer.
+
 ### Changed
 - Update nakama-common to v1.30.0 release.